Refine embedding bag kernels
