+++ src/aten/EmbeddingBag.cpp
@@ -1,11 +1,11 @@
 #include <ATen/ATen.h>
 #include <ATen/XPUNativeFunctions.h>
 
-#include <aten/sycl/EmbeddingBagKernel.h>
+#include <aten/sycl/EmbeddingBagKernels.h>
 
 namespace at {
 
-::std::tuple<Tensor, Tensor, Tensor, Tensor> XPUNativeFunctions::_embedding_bag(
+std::tuple<Tensor, Tensor, Tensor, Tensor> XPUNativeFunctions::_embedding_bag(
     const Tensor& weight,
     const Tensor& indices,
     const Tensor& offsets,
@@ -15,9 +15,25 @@ namespace at {
     const std::optional<Tensor>& per_sample_weights_opt,
     bool include_last_offset,
     int64_t padding_idx) {
+  TORCH_CHECK(
+      indices.dim() == 1 || indices.dim() == 2,
+      "input has to be a 1D or 2D Tensor, but got Tensor of dimension ",
+      indices.dim());
+  if (indices.dim() == 1) {
+    TORCH_CHECK(
+        offsets.dim() == 1,
+        "offsets has to be a 1D Tensor, but got Tensor of dimension ",
+        offsets.dim());
+  }
+  TORCH_CHECK(
+      weight.dim() == 2,
+      "weight has to be a 2D Tensor, but got Tensor of dimension ",
+      weight.dim());
+
   c10::MaybeOwned<Tensor> per_sample_weights_maybe_owned =
       at::borrow_from_optional_tensor(per_sample_weights_opt);
   const Tensor& per_sample_weights = *per_sample_weights_maybe_owned;
+
   return native::xpu::_embedding_bag_kernel(
       weight,
       indices,
@@ -30,7 +46,7 @@ namespace at {
       padding_idx);
 }
 
-::std::tuple<Tensor, Tensor, Tensor, Tensor> XPUNativeFunctions::
+std::tuple<Tensor, Tensor, Tensor, Tensor> XPUNativeFunctions::
     _embedding_bag_forward_only(
         const Tensor& weight,
         const Tensor& indices,
@@ -41,21 +57,15 @@ namespace at {
         const c10::optional<Tensor>& per_sample_weights_opt,
         bool include_last_offset,
         int64_t padding_idx) {
-  // See [Note: hacky wrapper removal for optional tensor]
-  c10::MaybeOwned<Tensor> per_sample_weights_maybe_owned =
-      at::borrow_from_optional_tensor(per_sample_weights_opt);
-  const Tensor& per_sample_weights = *per_sample_weights_maybe_owned;
-
-  return native::xpu::_embedding_bag_kernel(
+  return _embedding_bag(
       weight,
       indices,
       offsets,
       scale_grad_by_freq,
       mode,
       sparse,
-      per_sample_weights,
+      per_sample_weights_opt,
       include_last_offset,
       padding_idx);
 }
-
 } // namespace at
