+++ test/xpu/fin_grain/test_ops_xpu.py
@@ -136,6 +136,7 @@ _xpu_computation_op_list = [
     # "nn.functional.nll_loss", # Lack of XPU implementation of aten::nll_loss2d_forward. Will retrieve the case, only if the op is implemented.
     "sigmoid",
     "sgn",
+    "nn.functional.embedding_bag",
 ]
 
 _xpu_computation_ops = [
